SILA-3195 - Redesign Demo screen to avoid vertical scroll bar.
