+++ src/components/register/RegisterUserForm.js
@@ -9,7 +9,8 @@ import DefaultKYCForm from '../../components/register/DefaultKYCForm';
 import KYCLiteForm from '../../components/register/KYCLiteForm';
 import ReceiveOnlyKYCForm from '../../components/register/ReceiveOnlyKYCForm';
 import InstantAchKYCForm from '../../components/register/InstantAchKYCForm';
-import RegisterDataForm from '../../components/register/RegisterDataForm';
+import UpdateKYCForm from '../../components/register/UpdateKYCForm';
+import AddDataForm from '../../components/register/AddDataForm';
 
 import { DEFAULT_KYC, LITE_KYC, RECEIVE_ONLY_KYC, INSTANT_ACH_KYC, KYC_ARRAY } from '../../constants';
 
@@ -19,7 +20,8 @@ const RegisterUserForm = ({ className, handle, children, onError, onSuccess, onS
   const [errors, setErrors] = useState({});
   const [loaded, setLoaded] = useState(true);
   const [preferredKyc, setPreferredKyc] = useState(app.activeUser.kycLevel || app.settings.preferredKycLevel);
-  const [reloadUUID, setReloadUUID] = useState(false);
+  const [showUpdateBtn, setShowUpdateBtn] = useState(false);
+  const [entityuuid, setEntityuuid] = useState({ isFetchedUUID: false, uuid: {} });
 
   const register = async (e) => {
     console.log('\n*** BEGIN REGISTER USER ***');
@@ -43,35 +45,36 @@ const RegisterUserForm = ({ className, handle, children, onError, onSuccess, onS
       isValidated = false;
       validationErrors.entity = Object.assign({last_name: "This field may not be blank."}, validationErrors.entity);
     }
-    if (e.target.email && !e.target.email.value) {
+
+    if (preferredKyc !== RECEIVE_ONLY_KYC && e.target.email && !e.target.email.value) {
       isValidated = false;
       validationErrors.contact = Object.assign({email: "This field may not be blank."}, validationErrors.contact);
     }
-    if (e.target.phone && !e.target.phone.value) {
+    if (preferredKyc !== RECEIVE_ONLY_KYC && e.target.phone && !e.target.phone.value) {
       isValidated = false;
       validationErrors.contact = Object.assign({phone: "This field may not be blank."}, validationErrors.contact);
     }
-    if (e.target.dateOfBirth && !e.target.dateOfBirth.value) {
+    if (preferredKyc !== RECEIVE_ONLY_KYC && e.target.dateOfBirth && !e.target.dateOfBirth.value) {
       isValidated = false;
       validationErrors.entity = Object.assign({birthdate: "This field may not be blank."}, validationErrors.entity);
     }
-    if (e.target.ssn && !e.target.ssn.value) {
+    if (preferredKyc !== RECEIVE_ONLY_KYC && preferredKyc !== LITE_KYC && e.target.ssn && !e.target.ssn.value) {
       isValidated = false;
       validationErrors.identity = "This field may not be blank.";
     }
-    if (e.target.address && !e.target.address.value) {
+    if (preferredKyc !== RECEIVE_ONLY_KYC && preferredKyc !== LITE_KYC && e.target.address && !e.target.address.value) {
       isValidated = false;
       validationErrors.address = Object.assign({street_address_1: "This field may not be blank."}, validationErrors.address);
     }
-    if (e.target.city && !e.target.city.value) {
+    if (preferredKyc !== RECEIVE_ONLY_KYC && preferredKyc !== LITE_KYC && e.target.city && !e.target.city.value) {
       isValidated = false;
       validationErrors.address = Object.assign({city: "This field may not be blank."}, validationErrors.address);
     }
-    if (e.target.state && !e.target.state.value) {
+    if (preferredKyc !== RECEIVE_ONLY_KYC && preferredKyc !== LITE_KYC && e.target.state && !e.target.state.value) {
       isValidated = false;
       validationErrors.address = Object.assign({state: "This field may not be blank."}, validationErrors.address);
     }
-    if (e.target.zip && !e.target.zip.value) {
+    if (preferredKyc !== RECEIVE_ONLY_KYC && preferredKyc !== LITE_KYC && e.target.zip && !e.target.zip.value) {
       isValidated = false;
       validationErrors.address = Object.assign({postal_code: "This field may not be blank."}, validationErrors.address);
     }
@@ -85,7 +88,7 @@ const RegisterUserForm = ({ className, handle, children, onError, onSuccess, onS
     let ApiEndpoint;
     let updatedEntityData = {};
     let updatedResponses = [];
-    let successStatus = { entity: true, email: true, phone: true, identity: true, address: true };
+    let updateSuccess = false;
     if (app.activeUser && app.activeUser.handle) {
       let entityRes = {};
       try {
@@ -99,18 +102,18 @@ const RegisterUserForm = ({ className, handle, children, onError, onSuccess, onS
       }
 
       const entityUpdateData = {};
-      if (e.target.firstName.value !== app.activeUser.firstName) entityUpdateData.first_name = e.target.firstName.value;
-      if (e.target.lastName.value !== app.activeUser.lastName) entityUpdateData.last_name = e.target.lastName.value;
-      if (preferredKyc !== RECEIVE_ONLY_KYC && e.target.dateOfBirth.value !== app.activeUser.dateOfBirth) entityUpdateData.birthdate = e.target.dateOfBirth ? e.target.dateOfBirth.value : '';
+      if (e.target.firstName && e.target.firstName.value && e.target.firstName.value !== app.activeUser.firstName) entityUpdateData.first_name = e.target.firstName.value;
+      if (e.target.lastName && e.target.lastName.value && e.target.lastName.value !== app.activeUser.lastName) entityUpdateData.last_name = e.target.lastName.value;
+      if (e.target.dateOfBirth && e.target.dateOfBirth.value && e.target.dateOfBirth.value !== app.activeUser.dateOfBirth) entityUpdateData.birthdate = e.target.dateOfBirth ? e.target.dateOfBirth.value : '';
       if (Object.keys(entityUpdateData).length) {
         try {
           const entityUpdateRes = await api.updateEntity(app.activeUser.handle, app.activeUser.private_key, entityUpdateData);
           updatedResponses = [ ...updatedResponses, { endpoint: '/update/entity', result: JSON.stringify(entityUpdateRes, null, '\t') } ];
 
           if (entityUpdateRes.data.success) {
+            updateSuccess = true;
             updatedEntityData = { ...updatedEntityData, firstName: e.target.firstName.value, lastName: e.target.lastName.value, dateOfBirth: e.target.dateOfBirth.value};
           }  else if (entityUpdateRes.data.validation_details) {
-            successStatus = {...successStatus, entity: false};
             validationErrors = { ...validationErrors, entity: entityUpdateRes.data.validation_details }
           } else {
             console.log('... update entity failed!', entityUpdateRes);
@@ -121,138 +124,134 @@ const RegisterUserForm = ({ className, handle, children, onError, onSuccess, onS
         }
       }
 
-      if(preferredKyc !== RECEIVE_ONLY_KYC) {
-        validationErrors = { ...validationErrors, contact: {} }
-        if (e.target.email.value !== app.activeUser.email) {
-          try {
-            const emailUpdateData = {};
-            emailUpdateData.email = e.target.email ? e.target.email.value : '';
-            emailUpdateData.uuid = entityRes.data.emails[0] ? entityRes.data.emails[0]['uuid'] : '';
+      validationErrors = { ...validationErrors, contact: {} }
+      if (e.target.email && e.target.email.value && e.target.email.value !== app.activeUser.email) {
+        try {
+          const emailUpdateData = {};
+          emailUpdateData.email = e.target.email ? e.target.email.value : '';
+          emailUpdateData.uuid = entityRes.data.emails[0] ? entityRes.data.emails[0]['uuid'] : '';
 
-            ApiEndpoint = '/add/email';
-            let emailRes = {};
-            if (entityRes.data.emails.length) {
-              emailRes = await api.updateEmail(app.activeUser.handle, app.activeUser.private_key, emailUpdateData);
-              ApiEndpoint = '/update/email';
-            } else {
-              emailRes = await api.addEmail(app.activeUser.handle, app.activeUser.private_key, emailUpdateData.email);
-            }
-
-            updatedResponses = [ ...updatedResponses, { endpoint: ApiEndpoint, result: JSON.stringify(emailRes, null, '\t') } ];
-
-            if (emailRes.data.success) {
-              delete emailUpdateData.uuid;
-              updatedEntityData = { ...updatedEntityData, ...emailUpdateData }
-            } else if (emailRes.data.validation_details) {
-              successStatus = {...successStatus, email: false };
-              validationErrors.contact = Object.assign({email: emailRes.data.validation_details.email}, validationErrors.contact);
-            } else {
-              console.log('... update email failed!', emailRes);
-            }
-          } catch (err) {
-            console.log('  ... unable to update email, looks like we ran into an issue!');
-            handleError(err);
+          ApiEndpoint = '/add/email';
+          let emailRes = {};
+          if (entityRes.data.emails.length) {
+            emailRes = await api.updateEmail(app.activeUser.handle, app.activeUser.private_key, emailUpdateData);
+            ApiEndpoint = '/update/email';
+          } else {
+            emailRes = await api.addEmail(app.activeUser.handle, app.activeUser.private_key, emailUpdateData.email);
           }
-        }
 
-        const phoneUpdateData = {};
-        if (e.target.phone.value !== app.activeUser.phone) phoneUpdateData.phone = e.target.phone ? e.target.phone.value : '';
-        if (preferredKyc === INSTANT_ACH_KYC && e.target.smsOptIn && e.target.smsOptIn.checked !== app.activeUser.smsOptIn) phoneUpdateData.smsOptIn = (e.target.smsOptIn && e.target.smsOptIn.checked) ? true : false
-        if (Object.keys(phoneUpdateData).length) {
-          try {
-            ApiEndpoint = '/add/phone';
-            let phoneRes = {};
-            if (entityRes.data.phones.length) {
-              phoneUpdateData.phone = e.target.phone ? e.target.phone.value : '';
-              phoneUpdateData.uuid = entityRes.data.phones[0] ? entityRes.data.phones[0]['uuid'] : '';
-              phoneRes = await api.updatePhone(app.activeUser.handle, app.activeUser.private_key, phoneUpdateData);
-              ApiEndpoint = '/update/phone';
-            } else {
-              phoneRes = await api.addPhone(app.activeUser.handle, app.activeUser.private_key, phoneUpdateData.phone, {
-                smsOptIn: (e.target.smsOptIn && e.target.smsOptIn.checked) ? true : false
-              });
-            }
+          updatedResponses = [ ...updatedResponses, { endpoint: ApiEndpoint, result: JSON.stringify(emailRes, null, '\t') } ];
 
-            updatedResponses = [ ...updatedResponses, { endpoint: ApiEndpoint, result: JSON.stringify(phoneRes, null, '\t') } ];
-
-            if (phoneRes.data.success) {
-              updatedEntityData = { ...updatedEntityData, phone: e.target.phone.value }
-            } else if (phoneRes.data.validation_details) {
-              successStatus = {...successStatus, phone: false};
-              validationErrors.contact = Object.assign({phone: phoneRes.data.validation_details.phone}, validationErrors.contact);
-            } else {
-              console.log('... update phone failed!', phoneRes);
-            }
-          } catch (err) {
-            console.log('  ... unable to update phone, looks like we ran into an issue!');
-            handleError(err);
+          if (emailRes.data.success) {
+            delete emailUpdateData.uuid;
+            updateSuccess = true;
+            updatedEntityData = { ...updatedEntityData, ...emailUpdateData }
+          } else if (emailRes.data.validation_details) {
+            validationErrors.contact = Object.assign({email: emailRes.data.validation_details.email}, validationErrors.contact);
+          } else {
+            console.log('... update email failed!', emailRes);
           }
+        } catch (err) {
+          console.log('  ... unable to update email, looks like we ran into an issue!');
+          handleError(err);
         }
       }
 
-      if(preferredKyc === DEFAULT_KYC || preferredKyc === INSTANT_ACH_KYC) {
-        if (e.target.ssn.value !== app.activeUser.ssn) {
-          const identityUpdateData = {};
-          identityUpdateData.alias = 'SSN';
-          identityUpdateData.value = e.target.ssn ? e.target.ssn.value : '';
-          try {
-            ApiEndpoint = '/add/identity';
-            let ssnRes = {};
-            if (entityRes.data.identities.length) {
-              identityUpdateData.uuid = entityRes.data.identities[0] ? entityRes.data.identities[0]['uuid'] : '';
-              ssnRes = await api.updateIdentity(app.activeUser.handle, app.activeUser.private_key, identityUpdateData);
-              ApiEndpoint = '/update/identity';
-            } else {
-              ssnRes = await api.addIdentity(app.activeUser.handle, app.activeUser.private_key, identityUpdateData);
-            }
-
-            updatedResponses = [ ...updatedResponses, { endpoint: ApiEndpoint, result: JSON.stringify(ssnRes, null, '\t') } ];
-
-            if (ssnRes.data.success) {
-              updatedEntityData = { ...updatedEntityData, ssn: identityUpdateData.value }
-            } else if (ssnRes.data.validation_details) {
-              successStatus = {...successStatus, identity: false};
-              validationErrors = { ...validationErrors, identity: ssnRes.data.validation_details }
-            } else {
-              console.log('... update identity failed!', ssnRes);
-            }
-          } catch (err) {
-            console.log('  ... unable to update identity, looks like we ran into an issue!');
-            handleError(err);
+      const phoneUpdateData = {};
+      if (e.target.phone && e.target.phone.value && e.target.phone.value !== app.activeUser.phone) phoneUpdateData.phone = e.target.phone ? e.target.phone.value : '';
+      if (e.target.smsOptIn && e.target.smsOptIn.checked !== app.activeUser.smsOptIn) phoneUpdateData.smsOptIn = (e.target.smsOptIn && e.target.smsOptIn.checked) ? true : false
+      if (Object.keys(phoneUpdateData).length) {
+        try {
+          ApiEndpoint = '/add/phone';
+          let phoneRes = {};
+          if (entityRes.data.phones.length) {
+            phoneUpdateData.phone = e.target.phone ? e.target.phone.value : '';
+            phoneUpdateData.uuid = entityRes.data.phones[0] ? entityRes.data.phones[0]['uuid'] : '';
+            phoneRes = await api.updatePhone(app.activeUser.handle, app.activeUser.private_key, phoneUpdateData);
+            ApiEndpoint = '/update/phone';
+          } else {
+            phoneRes = await api.addPhone(app.activeUser.handle, app.activeUser.private_key, phoneUpdateData.phone, {
+              smsOptIn: (e.target.smsOptIn && e.target.smsOptIn.checked) ? true : false
+            });
           }
+
+          updatedResponses = [ ...updatedResponses, { endpoint: ApiEndpoint, result: JSON.stringify(phoneRes, null, '\t') } ];
+
+          if (phoneRes.data.success) {
+            updateSuccess = true;
+            updatedEntityData = { ...updatedEntityData, phone: e.target.phone.value }
+          } else if (phoneRes.data.validation_details) {
+            validationErrors.contact = Object.assign({phone: phoneRes.data.validation_details.phone}, validationErrors.contact);
+          } else {
+            console.log('... update phone failed!', phoneRes);
+          }
+        } catch (err) {
+          console.log('  ... unable to update phone, looks like we ran into an issue!');
+          handleError(err);
         }
+      }
 
-        const addressUpdateData = {};
-        if (e.target.address.value !== app.activeUser.address) addressUpdateData.street_address_1 = e.target.address ? e.target.address.value : '';
-        if (e.target.city.value !== app.activeUser.city) addressUpdateData.city = e.target.city ? e.target.city.value : '';
-        if (e.target.state.value !== app.activeUser.state) addressUpdateData.state = e.target.state ? e.target.state.value : '';
-        if (e.target.zip.value !== app.activeUser.zip) addressUpdateData.postal_code = e.target.zip ? e.target.zip.value : '';
-        if (Object.keys(addressUpdateData).length) {
-          try {
-            ApiEndpoint = '/add/address';
-            let addressRes = {};
-            if (entityRes.data.addresses.length) {
-              addressUpdateData.uuid = entityRes.data.addresses[0] ? entityRes.data.addresses[0]['uuid'] : '';
-              addressRes = await api.updateAddress(app.activeUser.handle, app.activeUser.private_key, addressUpdateData);
-              ApiEndpoint = '/update/address';
-            } else {
-              addressRes = await api.addAddress(app.activeUser.handle, app.activeUser.private_key, addressUpdateData);
-            }
-
-            updatedResponses = [ ...updatedResponses, { endpoint: ApiEndpoint, result: JSON.stringify(addressRes, null, '\t') } ];
-
-            if (addressRes.data.success) {
-              updatedEntityData = { ...updatedEntityData, address: e.target.address.value, city: e.target.city.value, state: e.target.state.value, zip: e.target.zip.value }
-            } else if (addressRes.data.validation_details) {
-              successStatus = {...successStatus, address: false};
-              validationErrors = { ...validationErrors, address: addressRes.data.validation_details.address }
-            } else {
-              console.log('... update address failed!', addressRes);
-            }
-          } catch (err) {
-            console.log('  ... unable to update address, looks like we ran into an issue!');
-            handleError(err);
+      if (e.target.ssn && e.target.ssn.value && e.target.ssn.value !== app.activeUser.ssn) {
+        const identityUpdateData = {};
+        identityUpdateData.alias = 'SSN';
+        identityUpdateData.value = e.target.ssn ? e.target.ssn.value : '';
+        try {
+          ApiEndpoint = '/add/identity';
+          let ssnRes = {};
+          if (entityRes.data.identities.length) {
+            identityUpdateData.uuid = entityRes.data.identities[0] ? entityRes.data.identities[0]['uuid'] : '';
+            ssnRes = await api.updateIdentity(app.activeUser.handle, app.activeUser.private_key, identityUpdateData);
+            ApiEndpoint = '/update/identity';
+          } else {
+            ssnRes = await api.addIdentity(app.activeUser.handle, app.activeUser.private_key, identityUpdateData);
           }
+
+          updatedResponses = [ ...updatedResponses, { endpoint: ApiEndpoint, result: JSON.stringify(ssnRes, null, '\t') } ];
+
+          if (ssnRes.data.success) {
+            updateSuccess = true;
+            updatedEntityData = { ...updatedEntityData, ssn: identityUpdateData.value }
+          } else if (ssnRes.data.validation_details) {
+            validationErrors = { ...validationErrors, identity: ssnRes.data.validation_details }
+          } else {
+            console.log('... update identity failed!', ssnRes);
+          }
+        } catch (err) {
+          console.log('  ... unable to update identity, looks like we ran into an issue!');
+          handleError(err);
+        }
+      }
+
+      const addressUpdateData = {};
+      if (e.target.address && e.target.address.value && e.target.address.value !== app.activeUser.address) addressUpdateData.street_address_1 = e.target.address ? e.target.address.value : '';
+      if (e.target.city && e.target.city.value && e.target.city.value !== app.activeUser.city) addressUpdateData.city = e.target.city ? e.target.city.value : '';
+      if (e.target.state && e.target.state.value && e.target.state.value !== app.activeUser.state) addressUpdateData.state = e.target.state ? e.target.state.value : '';
+      if (e.target.zip && e.target.zip.value && e.target.zip.value !== app.activeUser.zip) addressUpdateData.postal_code = e.target.zip ? e.target.zip.value : '';
+      if (Object.keys(addressUpdateData).length) {
+        try {
+          ApiEndpoint = '/add/address';
+          let addressRes = {};
+          if (entityRes.data.addresses.length) {
+            addressUpdateData.uuid = entityRes.data.addresses[0] ? entityRes.data.addresses[0]['uuid'] : '';
+            addressRes = await api.updateAddress(app.activeUser.handle, app.activeUser.private_key, addressUpdateData);
+            ApiEndpoint = '/update/address';
+          } else {
+            addressRes = await api.addAddress(app.activeUser.handle, app.activeUser.private_key, addressUpdateData);
+          }
+
+          updatedResponses = [ ...updatedResponses, { endpoint: ApiEndpoint, result: JSON.stringify(addressRes, null, '\t') } ];
+
+          if (addressRes.data.success) {
+            updateSuccess = true;
+            updatedEntityData = { ...updatedEntityData, address: e.target.address.value, city: e.target.city.value, state: e.target.state.value, zip: e.target.zip.value }
+          } else if (addressRes.data.validation_details) {
+            validationErrors = { ...validationErrors, address: addressRes.data.validation_details.address }
+          } else {
+            console.log('... update address failed!', addressRes);
+          }
+        } catch (err) {
+          console.log('  ... unable to update address, looks like we ran into an issue!');
+          handleError(err);
         }
       }
 
@@ -260,22 +259,6 @@ const RegisterUserForm = ({ className, handle, children, onError, onSuccess, onS
         console.log('  ... update completed!');
         let result = {};
         let appData = {};
-        let updateSuccess = false;
-
-        if (preferredKyc === DEFAULT_KYC || preferredKyc === INSTANT_ACH_KYC) {
-          if (successStatus.entity && successStatus.email && successStatus.phone && successStatus.identity && successStatus.address) {
-            updateSuccess = true;
-          }
-        } else if (preferredKyc === LITE_KYC) {
-          if (successStatus.entity && successStatus.email && successStatus.phone) {
-            updateSuccess = true;
-          }
-        } else if (preferredKyc === RECEIVE_ONLY_KYC) {
-          if (successStatus.entity) {
-            updateSuccess = true;
-          }
-        }
-
         if (updateSuccess) {
           refreshApp();
           const activeUser = app.users.find(u => u.handle === app.activeUser.handle);
@@ -287,8 +270,8 @@ const RegisterUserForm = ({ className, handle, children, onError, onSuccess, onS
           appData = {
             users: app.users.map(({ active, ...u }) => u.handle === app.activeUser.handle ? { ...u, ...updatedEntityData } : u),
           };
-          setReloadUUID(true);
           if (Object.keys(errors).length) setErrors({});
+          setShowUpdateBtn(false)
         } else if ( Object.keys(validationErrors).length ) {
           setErrors(validationErrors);
           setValidated(true);
@@ -375,7 +358,21 @@ const RegisterUserForm = ({ className, handle, children, onError, onSuccess, onS
     setPreferredKyc(value || undefined)
     setAppData({
       settings: { ...app.settings, preferredKycLevel: value || undefined }
+    }, () => {
+      if (app.activeUser && app.activeUser.handle) {
+        const activeUser = app.users.find(u => u.handle === app.activeUser.handle);
+        updateApp({ activeUser: app.activeUser ? { ...activeUser, kycLevel: value } : false });
+      }
     });
+  }
+
+  const updateUuid = (uuidObj) => {
+    setEntityuuid({...entityuuid, isFetchedUUID: true, uuid: {
+      email: uuidObj && uuidObj.email ? uuidObj.email : entityuuid.uuid.email ? entityuuid.uuid.email : '',
+      phone: uuidObj && uuidObj.phone ? uuidObj.phone : entityuuid.uuid.phone ? entityuuid.uuid.phone : '',
+      identity: uuidObj && uuidObj.identity ? uuidObj.identity : entityuuid.uuid.identity ? entityuuid.uuid.identity : '',
+      address: uuidObj && uuidObj.address ? uuidObj.address : entityuuid.uuid.address ? entityuuid.uuid.address : ''
+    } })
   }
 
   return (
@@ -391,12 +388,13 @@ const RegisterUserForm = ({ className, handle, children, onError, onSuccess, onS
         options={KYC_ARRAY} />
 
       {!preferredKyc ? <p className="text-right text-muted"><Button variant="link" className="text-reset font-italic p-0 text-decoration-none shadow-none" onClick={() => onShowKycModal(true)}><span className="lnk">What's the difference between these KYC levels?</span><i className="sila-icon sila-icon-info text-primary ml-2"></i></Button></p> : ''}
-      {preferredKyc === DEFAULT_KYC && <DefaultKYCForm errors={errors} app={app} isHide={(app.activeUser && app.activeUser.kycLevel === DEFAULT_KYC)} />}
-      {preferredKyc === LITE_KYC && <KYCLiteForm errors={errors} app={app} isHide={(app.activeUser && app.activeUser.kycLevel === LITE_KYC)} />}
-      {preferredKyc === RECEIVE_ONLY_KYC && <ReceiveOnlyKYCForm errors={errors} app={app} isHide={(app.activeUser && app.activeUser.kycLevel === RECEIVE_ONLY_KYC)} />}
-      {preferredKyc === INSTANT_ACH_KYC && <InstantAchKYCForm errors={errors} app={app} isHide={(app.activeUser && app.activeUser.kycLevel === INSTANT_ACH_KYC)} />}
-      {preferredKyc && (app.activeUser && app.activeUser.kycLevel !== preferredKyc) && <div className="d-flex mb-md-5"><Button type="submit" className="ml-auto" disabled={!(app.activeUser && app.activeUser.kycLevel !== app.settings.preferredKycLevel)}>Add data</Button></div>}
-      {app.activeUser && app.activeUser.handle && <RegisterDataForm errors={errors} onConfirm={onConfirm} onLoaded={(isLoaded) => setLoaded(isLoaded)} onErrors={(errorsObj) => { setErrors(errorsObj); setValidated(true); } } reloadUUID={reloadUUID} onReloadedUUID={(isReloaded) => setReloadUUID(isReloaded)} />}
+      {preferredKyc === DEFAULT_KYC && !app.activeUser && <DefaultKYCForm errors={errors} app={app} />}
+      {preferredKyc === LITE_KYC && !app.activeUser && <KYCLiteForm errors={errors} app={app} />}
+      {preferredKyc === RECEIVE_ONLY_KYC && !app.activeUser && <ReceiveOnlyKYCForm errors={errors} app={app} />}
+      {preferredKyc === INSTANT_ACH_KYC && !app.activeUser && <InstantAchKYCForm errors={errors} app={app} />}
+      {app.activeUser && app.activeUser.handle && <UpdateKYCForm errors={errors} preferredKyc={preferredKyc} entityuuid={entityuuid} onLoaded={(isLoaded) => setLoaded(isLoaded)} onConfirm={onConfirm} onShowUpdate={(isUpdated) => setShowUpdateBtn(isUpdated)} />}
+      {app.activeUser && app.activeUser.handle && <AddDataForm errors={errors} entityuuid={entityuuid} onLoaded={(isLoaded) => setLoaded(isLoaded)} onErrors={(errorsObj) => { setErrors(errorsObj); setValidated(true); } } onUpdateUuid={(uuidObj) => updateUuid(uuidObj)} />}
+      {preferredKyc && app.activeUser && (app.activeUser.kycLevel !== preferredKyc || showUpdateBtn) && <div className="d-flex mt-3 mb-md-3"><Button type="submit" className="ml-auto">Update data</Button></div>}
 
       {children}
     </Form>
